requirements.txt
psycopg2-binary==2.9.7
Django==3.2
django-environ==0.11.2
python-dotenv==1.0.0
